requirements.txt
torch>=2.1.0
torch_geometric
tqdm
numpy
transformer_lens
huggingface_hub
bitsandbytes
wandb
matplotlib
seaborn
scikit-learn
hydra-core
sae_lens
git+https://github.com/alessiodevoto/sparticles.git
tables
pandas
dspy